Add onnx as a python dependancy (#50)
Add onnx as a python dependancy using the "requirements.txt" and include "pip install -r requirements.txt" in the Makefile. This ensures that everytime the Python generator script is called it is ran with the versions specified in the requirements.txt file. No need have external dependacies in the third_party folder and no need to compile. The --onnx option is also removed from the Python script.